+++ src/main/java/net/dungeonz/mixin/PlayerEntityMixin.java
@@ -67,7 +67,7 @@ public abstract class PlayerEntityMixin extends LivingEntity {
     @Override
     public boolean addStatusEffect(StatusEffectInstance effect, Entity source) {
         if (!this.getWorld().isClient() && effect.getEffectType().value().isBeneficial() && this.getWorld().getRegistryKey() == DimensionInit.DUNGEON_WORLD) {
-            if (DungeonHelper.getDungeonPortalEntity((ServerPlayerEntity) (Object) this) != null && DungeonHelper.getDungeonPortalEntity((ServerPlayerEntity) (Object) this).getDisableEffects()) {
+            if (DungeonHelper.getCurrentDungeon((ServerPlayerEntity) (Object) this) != null && !DungeonHelper.getCurrentDungeon((ServerPlayerEntity) (Object) this).isPositiveEffectsAllowed()) {
                 return false;
             }
         }