+++ src/main/java/net/dungeonz/network/packet/DungeonPortalPacket.java
@@ -13,14 +13,15 @@ import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 
-public record DungeonPortalPacket(BlockPos blockPos, List<UUID> playerUuids, List<UUID> deadPlayerUuids, List<String> difficulties, Map<String, List<ItemStack>> possibleLoot,
+public record DungeonPortalPacket(String dungeonType, BlockPos blockPos, List<UUID> playerUuids, List<UUID> deadPlayerUuids, List<String> difficulties, Map<String, List<ItemStack>> possibleLoot,
                                   Map<String, List<ItemStack>> requiredItemStacks, int maxGroupSize, int minGroupSize, int waitingPlayerCount, int requiredLevel, int cooldownTime, String difficulty,
-                                  boolean disableEffects, boolean privateGroup, Optional<Identifier> backgroundId)
+                                  boolean allowEnderPearl, boolean allowPositiveEffects, boolean allowElytra, boolean allowRespawn, boolean privateGroup, Optional<Identifier> backgroundId)
         implements CustomPayload {
 
     public static final CustomPayload.Id<DungeonPortalPacket> PACKET_ID = new CustomPayload.Id<>(Identifier.of("dungeonz", "dungeon_portal_packet"));
 
     public static final PacketCodec<RegistryByteBuf, DungeonPortalPacket> PACKET_CODEC = PacketCodec.of((value, buf) -> {
+        buf.writeString(value.dungeonType);
         buf.writeBlockPos(value.blockPos);
         buf.writeCollection(value.playerUuids, (buffer, uuid) -> buffer.writeUuid(uuid));
         buf.writeCollection(value.deadPlayerUuids, (buffer, uuid) -> buffer.writeUuid(uuid));
@@ -33,14 +34,18 @@ public record DungeonPortalPacket(BlockPos blockPos, List<UUID> playerUuids, Lis
         buf.writeInt(value.requiredLevel);
         buf.writeInt(value.cooldownTime);
         buf.writeString(value.difficulty);
-        buf.writeBoolean(value.disableEffects);
+        buf.writeBoolean(value.allowEnderPearl);
+        buf.writeBoolean(value.allowPositiveEffects);
+        buf.writeBoolean(value.allowElytra);
+        buf.writeBoolean(value.allowRespawn);
         buf.writeBoolean(value.privateGroup);
         buf.writeOptional(value.backgroundId, PacketByteBuf::writeIdentifier);
 
-    }, buf -> new DungeonPortalPacket(buf.readBlockPos(), buf.readList((buffer) -> PacketByteBuf.readUuid(buffer)), buf.readList((buffer) -> PacketByteBuf.readUuid(buffer)),
+    }, buf -> new DungeonPortalPacket(buf.readString(), buf.readBlockPos(), buf.readList((buffer) -> PacketByteBuf.readUuid(buffer)), buf.readList((buffer) -> PacketByteBuf.readUuid(buffer)),
             buf.readList(PacketByteBuf::readString), buf.readMap(PacketByteBuf::readString, (bufx) -> ItemStack.LIST_PACKET_CODEC.decode(buf)),
             buf.readMap(PacketByteBuf::readString, (bufx) -> ItemStack.LIST_PACKET_CODEC.decode(buf)), buf.readInt(),
-            buf.readInt(), buf.readInt(), buf.readInt(), buf.readInt(), buf.readString(), buf.readBoolean(), buf.readBoolean(), buf.readOptional(PacketByteBuf::readIdentifier)));
+            buf.readInt(), buf.readInt(), buf.readInt(), buf.readInt(), buf.readString(), buf.readBoolean(), buf.readBoolean(), buf.readBoolean(), buf.readBoolean(), buf.readBoolean(),
+            buf.readOptional(PacketByteBuf::readIdentifier)));
 
     @Override
     public Id<? extends CustomPayload> getId() {