+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ### Added:
-- 
+- Added ender pearl dungeon option
+- Added info button
 ### Fixed:
 - Fixed open screen in other dimensions than overworld
 - Fixed lever placing in dungeons
 - Fixed keep inventory option
 - Fixed water removing portal
 ### Changed:
-- 
+- Changed positive effects to data option